Organizing the tuist configuration
