+++ README.md
@@ -34,6 +34,19 @@ Download the dmg - [click here](https://github.com/heroesofcode/swift-explorer/r
 
 Please see the [documentation](https://github.com/heroesofcode/swift-explorer/blob/main/CONTRIBUTING.md)
 
+## Run the project
+
+Install [mise](https://mise.jdx.dev/)
+```sh
+mise install
+```
+
+Generate `SwiftExplorer.xcworkspace` with [Tuist](https://tuist.io/)
+```
+tuist install
+tuist generate
+```
+
 ## Inspiration
 
 This project was inspired by [kotlin-explorer](https://github.com/romainguy/kotlin-explorer) developed by [Romain Guy](https://github.com/romainguy).
