+++ Tuist/Config.swift
@@ -1,9 +1,3 @@
 import ProjectDescription
 
-let config = Config(
-//    Create an account with "tuist auth" and a project with "tuist project create"
-//    then uncomment the section below and set the project full-handle.
-//    * Read more: https://docs.tuist.io/guides/quick-start/gather-insights
-//
-//    fullHandle: "{account_handle}/{project_handle}",
-)
+let config = Config()
