+++ Tuist/Package.swift
@@ -4,12 +4,7 @@ import PackageDescription
 #if TUIST
     import ProjectDescription
 
-    let packageSettings = PackageSettings(
-        // Customize the product types for specific package product
-        // Default is .staticFramework
-        // productTypes: ["Alamofire": .framework,] 
-        productTypes: [:]
-    )
+    let packageSettings = PackageSettings(productTypes: [:])
 #endif
 
 let package = Package(
