+++ Tuist/ProjectDescriptionHelpers/Script.swift
@@ -1,0 +1,18 @@
+import ProjectDescription
+
+public extension Project {
+
+    static func targetScripts() -> [TargetScript] {
+        return [
+            TargetScript.pre(
+                script:
+                """
+                export PATH="$HOME/.local/share/mise/shims:$PATH"
+                swiftgen
+                """,
+                name: "SwiftGen",
+                basedOnDependencyAnalysis: false
+            )
+        ]
+    }
+}
